+++ Applicationv.01/admin/majordesription.aspx.cs
@@ -11,10 +11,12 @@ public partial class admin_majordesription : System.Web.UI.Page
     Logger objLog = new Logger();
     private GTEEntities db = new GTEEntities();
     string webURL = System.Configuration.ConfigurationManager.AppSettings["WebUrl"].ToString();
+    int universityID;
     protected void Page_Load(object sender, EventArgs e)
     {
         if (!Utility.CheckAdminLogin())
             Response.Redirect(webURL + "admin/Login.aspx", true);
+        universityID = Convert.ToInt32(System.Configuration.ConfigurationManager.AppSettings["UniversityID"].ToString());
         if (!IsPostBack)
             BindMajorDescription();
     }
@@ -60,12 +62,22 @@ public partial class admin_majordesription : System.Web.UI.Page
             {
                 majordiscipline_master objMajor = new majordiscipline_master();
 
-                TextBox txtDescription = (TextBox)gvMajorDescription.FooterRow.FindControl("txtDescription1");                
+                TextBox txtDescription = (TextBox)gvMajorDescription.FooterRow.FindControl("txtDescription1");
 
-                objMajor.description = txtDescription.Text.Trim();
-                db.majordiscipline_master.Add(objMajor);
-                db.SaveChanges();
-                BindMajorDescription();
+                var existingData = (from data in db.majordiscipline_master
+                                    where data.description == txtDescription.Text.Trim()
+                                    select data.description).FirstOrDefault();
+                if (string.IsNullOrEmpty(existingData))
+                {
+                    objMajor.description = txtDescription.Text.Trim();
+                    db.majordiscipline_master.Add(objMajor);
+                    db.SaveChanges();
+                    BindMajorDescription();
+                }
+                else
+                {
+                    ScriptManager.RegisterClientScriptBlock(this, this.GetType(), "alertMessage", "alert('Data is already recorded with entered description ')", true);
+                }
             }
         }
         catch (Exception ex)
@@ -114,7 +126,11 @@ public partial class admin_majordesription : System.Web.UI.Page
             var existsInApplicantdetails = db.applicantdetails.Where(c => c.fieldofhigheststudy == ID).ToList();
             var existsIngteapplicantdetails = db.gte_applicantdetails.Where(g => g.highestqualificationfield == ID && g.fieldofstudyapplied == ID).ToList();
             var existsIncoursemaster = db.coursemaster.Where(c => c.modeofstudyId == ID).ToList();
-            if (existsInApplicationmaster.Count == 0 && existsInApplicantdetails.Count == 0 && existsIngteapplicantdetails.Count == 0 && existsIncoursemaster.Count == 0)
+            var existsInUniversitywisemapping = (from umm in db.universitywisemastermapping
+                                                 join mn in db.master_name on umm.masterid equals mn.masterid
+                                                 where umm.universityid == universityID && mn.mastername.ToUpper().Contains("Major Discipline Master") && umm.mastervalueid == ID
+                                                 select umm).ToList();
+            if (existsInApplicationmaster.Count == 0 && existsInApplicantdetails.Count == 0 && existsIngteapplicantdetails.Count == 0 && existsIncoursemaster.Count == 0 && existsInUniversitywisemapping.Count == 0)
             {
                 db.majordiscipline_master.Remove(objMajor);
                 db.SaveChanges();
@@ -172,6 +188,29 @@ public partial class admin_majordesription : System.Web.UI.Page
     protected void gvMajorDescription_PageIndexChanging(object sender, GridViewPageEventArgs e)
     {
         gvMajorDescription.PageIndex = e.NewPageIndex;
+        BindMajorDescription();
+    }
+
+    protected void Add(object sender, EventArgs e)
+    {
+        Control control = null;
+        if (gvMajorDescription.FooterRow != null)
+            control = gvMajorDescription.FooterRow;
+        else
+            control = gvMajorDescription.Controls[0].Controls[0];
+        string idDescriptonText = (control.FindControl("txtEmptyRecordDescription") as TextBox).Text;
+        if (string.IsNullOrEmpty(idDescriptonText))
+        {
+            ScriptManager.RegisterClientScriptBlock(this, this.GetType(), "alertMessage", "alert('Description Cannot Be Empty')", true);
+            return;
+        }
+
+        majordiscipline_master objID = new majordiscipline_master();
+
+        objID.description = idDescriptonText;
+
+        db.majordiscipline_master.Add(objID);
+        db.SaveChanges();
         BindMajorDescription();
     }
 }
